Prepare for splitting marriage-abroad outcome for BOT
BOT = British Overseas Territories

The current outcome for ceremonies in BOT
(outcome_opposite_sex_marriage_in_british_overseas_territory) contains three
logical outcomes:

* One for ceremonies in British Indian Ocean Territory
* One for ceremonies in British Virgin Islands
* One for ceremonies in other British Overseas Territories

The outcome body contains a single `if/else` statement that decides which
logical outcome to display. Splitting these logical outcomes and moving the
"routing logic" to the flow is more consistent with the majority of this Smart
Answer.

This change will make it easier to split the outcome in a subsequent commit.