Regenerate regression tests artefacts (SFF)
This commit shows the affected/changes to regression test artefacts as a result of the Content
changes to the file sizes from the commit before this.

I have reviewed these changes and found it to be as expected.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no effect.