Update Gaza procedure to apply for an overseas passport.
A user need to call the Visa Application Centre to apply for a new
passport (or renew, apply, replace). The passport will be delivered to
the UK Visa Application Centre where the user applied.

Unfortunately the Uk Visa Application Centre in Gaza is not listed within
its list of organisations. Unfortunately, due to a deadline for this
Trello card [1] the address is inserted in the template, but the ideal
approach would be to retrieve this data from the same service as the
other organisations.

I a not also extracting this address into a partial, nor YML file until
we have a clear picture of how to deal with this situation.

There are two list of UK Visa Application Centres, I am not refactoring
them as part of this commit as I wan to keep a small scope.

[1]: https://trello.com/c/qPAaFIkL/195-13-july-overseas-passports-courier-fees-and-address-change-content-change-request