Add a YAML file with the features of a home
The heading for the boilers and insulation section is displayed even if
the user has selected all possible features for their type of home, and
there will therefore be no bullet points to display under the heading.

There are currently three home types:

modern
older
historic

Each home type has a different list of possible features. When the
outcome is rendered, there is no easy way to determine the total number
of features the user was presented with to select from.

The new YAML date file contains a list of all of the possible features
each type of home could have.

The list will be used to determine whether the user has selected all of
the features for their home type.

If they have selected all of the possible features, then the boilers and
insulation heading won't be displayed on the outcome.