Update condition for rendering boilers template
The heading for the boilers and insulation section is displayed even if
the user has selected all possible features for their type of home, and
there will therefore be no bullet points to display under the heading.

The energy grants calculator flow now uses the new convenience methods
in the model:
home_features_modern
home_features_older
home_features_historic

to compare the number of features a property has to the total number of
features the property could have.

This new state variable is passed to the outcome templates and used in
the conditions that determine whether the section on boilers and
insulation will be displayed.