Refactor home features questions in the flow
The options used in the home features questions are also being
defined in the YAML. To avoid duplication, and the possibility of the
two lists going out of whack, it seems sensible to only define the
options in one place.

Refactor the home features questions in the flow to read in their
options list from the energy grants calculator YAML file.