Make registration_country use current_country or country_of_death
If the `current_country` attribute has a value then it should be used to
determine `registration_country` in preference to `country_of_birth`.

Previously this behaviour was being provided by the re-defining of a
`registration_country` calculate block in the `which_country_are_you_in_now?`
question, a calculate block which was previously defined in `which_country?`.

This mirrors the way I implemented this behaviour in the new-style version of
the `register-a-birth` flow.