Don't raise InvalidResponse in overseas_passports_embassies method
It doesn't make any sense to me to raise this validation-related exception in an
outcome, because the DSL isn't intended to be used in this way. What's more
`WorldLocation.find` should typically work for all countries in the two
`country_select` questions. If it doesn't I'm sure we'd quickly see a useful
exception pointing to the problem.