Split courier info for IPS 3 online country
As per our discussion with Ben Mortimer. This brings the display of this
information inline with that for Tajikistan, Turkmenistan and
Uzbekistan. This should help us remove duplication in future.

We don't know what defines IPS3 countries. This just seemed like a
natural way to split the changes in the branch. We're happy to accept
the changes to the regression test artefacts because they match the
change we made to the template.

NOTE. No regression tests failed after making this change. We've done
some initial investigation and think that this path of code might be
unreachable. We're not going to investigate any further at the moment as
our primary goal is to make the content consistent so that we can remove
duplication.