Split courier info for replacing passport in IPS 1 country
As per our discussion with Ben Mortimer. This brings the display of this
information inline with that for Tajikistan, Turkmenistan and
Uzbekistan. This should help us remove duplication in future.

We don't know what defines IPS1 countries. This just seemed like a
natural way to split the changes in the branch. We're happy to accept
the changes to the regression test artefacts because they match the
change we made to the template.