Somalia can't register a birth with UK authorities
By including somalia within the list of countries with no embassy, the
flow goes straight to the `no_embassy_result.govspeak.erb` outcome.
In that outcome it is explained to the user that he needs to register
the birth according to the regulations in Somalia.

We are getting many deletions in the artefacts because in the
question `country_of_birth?` we jump straight to the outcome
`no_embassy_result`.