Somalia can't register a death with UK authorities
By including somalia within the list of countries with no embassy, the
flow goes straight to the `no_embassy_result.govspeak.erb` outcome. In
that outcome it is explained to the user that he needs to register the
death according to the regulations in Somalia.

We are getting many deletions in the artefacts because in the question
`which_country?` we jump straight to the outcome `no_embassy_result`.