Use pay_method calculator attribute instead of pay_pattern state variable
I want to avoid explicitly calculating the average weekly pay i.e. I want
methods that need the average weekly pay to trigger the calculation
automatically.

In order to do that I will need both the arguments to the
`calculate_average_weekly_pay` method to be available as calculator attributes.

I noticed that `pay_pattern` was already being stored in the `pay_method`
calculator attribute and so this commit is a small step towards what I want to
achieve.

Note that I added a `calculate` block to the `pay_frequency?` question in
`MaternityCalculatorFlow` to match the code in the other two sub-flows.