Store earnings_for_pay_period as attribute on calculator
I want to avoid explicitly calculating the average weekly pay i.e. I want
methods that need the average weekly pay to trigger the calculation
automatically.

In order to do that I will need both the remaining argument to the
`calculate_average_weekly_pay` method to be available as a calculator attribute.