Add response for date of birth of 1st May 1953
The content on the outcome pages where the user has entered a birth
date between 5th July 1951 and 5th May 1953 should change when the new
winter fuel payments threshold date is applied.

Added a new response for 1st May 1953 which is within this date range
so that the changes in the artefacts can be highlighted.