Add rates file for winter fuel payment threshold
The threshold date for winter fuel payments changes every year.
It would be more efficient to treat them as an uprating and make
changing the date just a matter of updating a YAML file.

Set start and end dates to the current (2016-201)7 tax year

We don't actually know how the long the winter fuel threshold date is
valid for, but as the latest available threshold date will be picked up
even if a new threshold is not added by the end of the tax year, it
seems safe enough to set an end date for the threshold in the near
future.