Regenerate test artefacts energy-grants-calculator
The test artefacts have been updated as expected when the date of birth
entered is 1st May 1953.

All of the outcomes for this date of birth now correctly show the text
about winter fuel payments

There are a few other interesting changes.

1. If you are eligible for winter fuel payment, then you are no longer
eligible_for_energy_company_obligation. This means that the text to
"get help with insulation or boiler repairs" is removed from the
relevant templates.

2. In one outcome, cold weather payment text will render if you are
eligible for winter fuel payment.

In the other two templates where this text could be rendered, the user
has to be eligible for cold weather payment.

So in some artefacts, but not all, the cold weather payment text will
suddenly appear.

These anomalies are being addressed in story:
https://trello.com/c/fwdVopYW