Generate artefacts for Somalia in Register a Death
This is a preparatory commit that generate all artefacts for Somalia for
register a death. The goal is to be sure in incoming commits that all the
incoming changes are expected.

I am also regenerating Register a Death because, register a birth and
register a death share the file `registration.yml` so both tools will
be affected in the incoming commits.