Remove registration mapping from Somalia to Ethiopia
The registration country for Somalia is no longer Ehtiopia. This
mapping were previously stored in the file `registrations.yml` which is
used in `RegistrationsDataQuery.rb`

This helper file is being used in:
- register-a-birth
- register-a-death
- marriage-abroad

We get a change from Ethiopia to Somalia in Register a birth as expected.

We are not getting the same change in register-a-death because we are
using a different method to retrieve the mapped country in both
calculators. In register a birth we are using:
`registration_country_name_lowercase_prefix` and in register-a-death we
are using: `death_country_name_lowercase_prefix`
It is not completely clear why a different method is being used, but as
we are getting the expected results described in [2], we postpone
investigating further.

And I am confident that marriage-abroad is not affected by this change because
I have generated all artefacts for all countries and after deleting the
mapping I had no failing tests.

[1]: https://trello.com/c/EX4RIdqc/217-change-to-registering-a-birth-where-child-is-born-in-somalia-content-change-request
[2]: https://trello.com/c/ssuabCG9/216-change-to-registering-a-death-in-somalia-content-change-request