Use earnings_for_pay_period instead of average_weekly_earnings
These two uses of the `average_weekly_earnings` calculator attribute in `if`
conditions were relying on the fact that it would return `nil` (i.e. false-y) if
the `calculate_average_weekly_pay` calculator method had not been called.

I'm planning to remove the `calculate_average_weekly_pay` method and convert the
`average_weekly_earnings` attribute into a fully-fledged *method* which will do
the calculation *on demand*. In the case where `earnings_for_pay_period` has not
been set, this new `average_weekly_earnings` method would return `0.0`, because
`nil.to_f` is `0.0`.

This in turn would mean that the following condition would be `true`:

    calculator.average_weekly_earnings < calculator.lower_earning_limit

However, as far as I can tell from the tests, the intention is that this
condition should be `false` if the `earnings_for_pay_period` has not been set.

So I've made this more explicit by checking `earnings_for_pay_period` instead
of `average_weekly_earnings` in the relevant `if` conditions. This should help
with the planned change described above.

Although I haven't investigated thoroughly, it looks to me as if the `if`
condition [1] in the outcome template is always `false`. If this is correct,
then these lines of content [2] will never be displayed. If this behaviour is
correct, then the condition and content should be removed. If it is not correct
then it should be fixed. I've created a new Trello card [3] to investigate this.

[1]: https://github.com/alphagov/smart-answers/blob/d5204882b8f7d9bd876ce5e362442e7e0f5c0a46/lib/smart_answer_flows/maternity-paternity-calculator/outcomes/maternity_leave_and_pay_result.govspeak.erb#L24
[2]: https://github.com/alphagov/smart-answers/blob/d5204882b8f7d9bd876ce5e362442e7e0f5c0a46/lib/smart_answer_flows/maternity-paternity-calculator/outcomes/maternity_leave_and_pay_result.govspeak.erb#L25-L29
[3]: https://trello.com/c/sFn7cp8B