Convert average_weekly_earnings into fully-fledged calculator method
It's simpler to calculate `average_weekly_earnings` on demand, rather than
having to remember to explicitly call `calculate_average_weekly_pay` beforehand.

This commit converts the `average_weekly_earnings` attribute reader method into
a fully-fledged calculator method, using the logic from the
`calculate_average_weekly_pay` method. It also removes calls to the
`calculate_average_weekly_pay` method as well as the definition of that method.

Note that it appears that nothing was calling the `average_weekly_earnings`
attribute setter method, so I chosen not to retain it.

Note also that the `next_node_calculation` & `calculate` blocks in which
`calculate_average_weekly_pay` was being called, were technically also setting
the `calculator` state variable. However, the new value (a reference to the
instance of the calculator) was always the same as the old one. Thus this aspect
of their behaviour was probably only there because you are required to provide
a return value for these types of block. Once the calls to
`calculate_average_weekly_pay` were removed, the need for the blocks disappeared
and so I have removed them.

Ideally I wouldn't have used memo-ization in the new `average_weekly_earnings`
method, because it's a premature optimisation. However, using it makes the
behaviour closer to the original behaviour of the method in that the value is
only calculated once. Also it turns out that one of the integration tests [1]
fails if this memo-ization is not in place. Unfortunately the test name and the
assertion failure message [2] are not very illuminating, so for now I haven't
investigated any further and have just used the memo-ization:

    MaternityCalculatorTest#test_: answer maternity given the date is April 9th (after changes) answer 21 November 2012 no contract answer 21 November 2012 answer yes answer yes answer 2 days before Saturday of qualifying week answer 8 weeks before qualifying week answer every month answer 1807.78 should calculate the dates and payment amounts.  [test/integration/smart_answer_flows/maternity_calculator_test.rb:485]:
    Expected: "187.73"
      Actual: "203.38"

[1]: https://github.com/alphagov/smart-answers/blob/d77f4c91de6de5ce6a1bca2eb594ccf51bc74cbd/test/integration/smart_answer_flows/maternity_calculator_test.rb#L485