Convert average_weekly_earnings into fully-fledged calculator method
It's simpler to calculate `average_weekly_earnings` on demand, rather than
having to remember to explicitly call `calculate_average_weekly_pay` beforehand.

This commit converts the `average_weekly_earnings` attribute reader method into
a fully-fledged calculator method, using the logic from the
`calculate_average_weekly_pay` method. It also removes calls to the
`calculate_average_weekly_pay` method as well as the definition of that method.

I chose not to memo-ise the result of the `average_weekly_earnings` in an
instance variable, because the logic in the method doesn't seem complex enough
to be worthy of potentially premature optimisation.

Note that it appears that nothing was calling the `average_weekly_earnings`
attribute setter method, so I chosen not to retain it.

Note also that the `next_node_calculation` & `calculate` blocks in which
`calculate_average_weekly_pay` was being called, were technically also setting
the `calculator` state variable. However, the new value (a reference to the
instance of the calculator) was always the same as the old one. Thus this aspect
of their behaviour was probably only there because you are required to provide
a return value for these types of block. Once the calls to
`calculate_average_weekly_pay` were removed, the need for the blocks disappeared
and so I have removed them.