Store pay_pattern as attribute on calculator
I want the calculator to be able to calculate average weekly earnings without
the flow having to explicitly call the `calculate_average_weekly_pay` method. In
order to do this the calculator will need access to the values of the two
arguments currently being passed in, `pay_pattern` & `pay`.

This commit makes a small step in that direction by storing the `pay_pattern`
response in a new calculator attribute, `pay_pattern`. I considered using the
existing `pay_method` attribute, but that gets over-written later in the flow.

The new `pay_pattern` attribute replaces the `pay_pattern` state variable, which
is why I was able to remove the `calculate` block from
`MaternityPaternityCalculatorFlow` and the `save_input_as` blocks from the sub-
flows.