Convert next_node_calculation blocks into on_response blocks
This is part of an effort to move all flows towards a new style [1]. In the new
style, `next_node_calculation` blocks are deprecated. This commit removes the
only two remaining usages in production code (once other pending PRs have been
merged).

Note that I've chosen to update the regression test checksums in this commit,
because all the regression tests are still passing i.e. this was just an
internal refactoring; no external behaviour has changed.

[1]: https://github.com/alphagov/smart-answers/blob/master/doc/refactoring.md