Add unit tests for under_green_deal?
The function that determines whether a user is eligible for the green
deal is very difficult to read.

As a starting point, I've added some unit tests to determine the cases
when a user is eligible for the green deal.

This function should be refactored to make it easier to update the
eligibilty rules in the future if needed, however more tests will have
to be added to cover all of the eligibility criteria before this can
even be attempted.