Remove unused `country_question` partial
This doesn't seem to be used anywhere in the code. I searched the code
base for "country_question" and couldn't find any reference to it
(except to "country_question_title"). The original logic behind it was
removed as part of the Smartdown code in #2050. Although I assume
`country_select_question` is its substitute, it's difficult to find
any evidence for that.