Remove unused `text_question` partial
This doesn't seem to be used anywhere in the code. I searched the code
base for "text_question" and couldn't find any reference to it.
The original logic behind it was removed as part of the Smartdown code
in #2050. Although I assume `value_question` is its substitute, it's
difficult to find any evidence for that.