Simplify conditional logic around link to consular fees
The commit titled "Display payment information independently to consular fee
link" means that I'm now able to have this single conditional that determines
whether to display the link to consular fees. Prior to this commit, the outer
conditional decided whether we can potentially display both the link to
consular fees and payment information partial.

Note that the logic has been inverted as it's changed from being an `if`
condition to an `unless` condition.