Add a "Deploy to Heroku" button to the README
The URL for this button does not explicitly specify a template to use. This
means Heroku will use the `Referer` HTTP header to determine the fork and branch
to deploy. The idea being that the button should work on the README page of
*any* fork and/or branch of the repo.

See this Heroku documentation [1] for more info.

[1]: https://devcenter.heroku.com/articles/heroku-button