Enlarge hit area of checkboxes
This gives checkbox questions the standard design
and makes it consistent with multiple choice questions.
It increases the checkboxes' hit area so that it's easier
to hit the otherwise tiny boxes.

We use styles from the `static` repo for this which are undocumented,
but the pattern (not the HTML/CSS) is the same as in Elements:
* http://govuk-elements.herokuapp.com/form-elements/#form-radio-buttons
* http://govuk-elements.herokuapp.com/form-elements/#form-checkboxes

Any flow with a checkbox question will be affected:

* benefit-cap-calculator
* calculate-statutory-sick-pay
* energy-grants-calculator
* legalisation-document-checker
* maternity-paternity-calculator
* simplified-expenses-checker
* student-finance-calculator