Introduce QuestionWithOptionsPresenter into inheritance chain
Only multiple choice & checkbox questions need the `QuestionPresenter#options`
method. This method doesn't work for other question types, because the
underlying question node doesn't have an `options` method.

I'm planning on pushing the current implementation of
`QuestionPresenter#options` down onto `QuestionWithOptionsPresenter` and
replacing it with an implementation which always returns an empty array. That
way it will always be safe to call the method.