Push QuestionPresenter#options down -> QuestionWithOptionsPresenter
Previously calling `QuestionPresenter#options` for questions with no options
would cause an error. This commit pushes the original implementation down onto
`QuestionWithOptionsPresenter` and replaces it with an implementation which
always returns an empty Array. This means that it should now always be safe to
call the `#options` method on the presenter for any question type.