Add AB testing setup for July 2016 session
This commit sets up AB test for the following smart answers:

* overseas-passport
* calculate-your-child-maintenance
* marriage-abroad

Cohorts across the aforementioned smart answers fall under the following.

* Start now
* Continue
* Next

The isStartPage checks that the current path matches the slug of the smart answers that has been identified for A/B testing.

It splits the pathname with the forward slash delimiter and joins all.

This in effect returns a combination of all the characters in the path for the current page, returning true for only paths that equal the defined slug.