Fix question layout when increasing font size
The question page breaks when the font size gets increased,
making the content too narrow, showing only 1-2 words per line.
This is because the grid of the question page does not use
the standard grid and it's using `em` instead of `px` or `%`.
This fix does not implement the standard grid as that would mean
changing all 396 question artefacts. But it does the next best thing
which design-wise and behaviour-wise is the same as our standard grid
but code-wise it is not. This also includes changing the width
of the question page content from 7/12 to 2/3, which is our standard.