Remove redundant regression test artefacts for calculate-state-pension
This flow was renamed to `state-pension-age` in this enormous commit [1].
Unfortunately, while the relevant files in `test/data/calculate-state-pension`
were moved to `test/data/state-pension-age`, the files in
`test/artefacts/calculate-state-pension` were left in place and new ones were
generated in `test/artefacts/state-pension-age`.

Since that commit, the artefacts removed in this commit have not been used and
can safely be removed.

[1]: 1883db4f1a83b17b71074308bcd3248624faf021