Fix commented out code in flow_preview initializer
This commented out code is intended to allow a developer to use the app to view
fixture flows. Since the code is commented out, inevitably it has become
out-of-date. I'm not totally convinced it's worth keeping this hanging around,
but in the meantime I thought it was worth fixing it.