Add content_id for all-smart-answer-questions draft flow
Even draft flows must have a `content_id`, because they are registered with the
Publishing API in the integration environment.

I've added a test to catch this problem, but it would be better to modify the
`Flow` class to effectively make it impossible to create a flow without a
`content_id`.