Add function to annotate content with link to template on Github
This JavaScript relies on the `debug-template-path` data attributes made
available in #2667. This idea is to call the new function from a bookmarklet or
the GOV.UK Chrome extension.

I wanted to add the new function in a separate file, but I couldn't get this to
work. So for the moment I've added it in the existing `smart-answers.js` file.

Ideally I would've added some unit tests around the function. However, we don't
have any such tests at the moment, so it would be a fair bit of work.

Today is my last day working on Smart Answers, I wanted to get this code live so
the Content Designers could start using it. Since it's not critical
functionality I think not having tests is OK for the moment.