Regenerate test artefacts energy-grants-calculator
The outcomes that displayed solid wall text if only main gas wasn’t
selected, or did not have a condition around the solid wall text at all,
have had the solid wall text removed.
    
Both solid wall and main gas have to not be selected for the text to
appear.