Introduce PartialTemplateWrapper class
This class is intended to be used as a handler for the
`PartialTemplateRenderInterceptor` class in order to wrap the content rendered
from partial templates with markup indicating the path to the source template.

Currently the rendered content is wrapped in a `div` element with a data
attribute, `data-debug-partial-template-path`, whose value is set to the path to
the source template relative to the root directory of the app.

A "markdown" attribute is needed in order to make govspeak (actually kramdown)
process the govspeak inside the `div`, because this won't happen by default [1].
I plan to add an "integration" test in a later commit which will ensure that
this attribute has the desired effect.

[1]: http://kramdown.gettalong.org/syntax.html#html-blocks