Annotate content rendered from flow partials w/ path to template
In #2667 we started annotating the content rendered from flow templates with the
path to the source template via a data attribute, `debug-template-path`.

This commit annotates the content rendered from flow *partial* templates with
the path to the source template via a similar data attribute,
`debug-partial-template-path`. The plan is to use it in a JavaScript function
like the one added in #2670 so that Content Designers can easily find the
template relating to specific content in the GitHub repository.

In order to achieve this, the `PartialTemplateWrapper` is wired up to
`ActionView::PartialRenderer` using `PartialTemplateRenderInterceptor` which
were introduced in earlier commits.

I've added a Rails functional test to act as an "integration" test to check
everything is wired up correctly. In particular this should protect against
breakages due to changes in `ActionView::PartialRenderer` and changes in
govspeak or kramdown.

There didn't seem much point in including the annotation in the regression test
artefacts and so I've introduced an environment variable to disable the
annotation in that case. This also has the benefit that I haven't had to update
all the regression test artefacts which were generated using one or more partial
templates.