Fix question layout when increasing font size
The question page breaks when the font size gets increased,
making the content too narrow, showing only 1-2 words per line.
This is because the grid of the question page does not use
the standard grid and it's using `em` instead of `px` or `%`.

This implements the standard grid by using the `grid-row` class
in a container around *all* page types,
making them consistent with mainstream.
Because of that I also had to remove the hack for question pages
(in `.step.current`) and an alternative version non-standard version
of the grid-row for outcome pages (`article.outcome`).

Side effects of this are:

* the width of question page content changes from 7/12
  to our standard of 2/3
* the 2/3 width of question and outcome page content
  now includes the h1, previously the h1 was shown in 100% width