Render payment_information in outcome for Brazil
This brings the template into line with the os_affirmation/fees partial, which
will make it easier to switch to in a subsequent commit.

This addition doesn't change the behaviour. The payment_information partial
only renders content if the country has a `payment_partial_name` set in
marriage_abroad_services.yml which Brazil doesn't.