Always render consular_fees_table partial
The consular_fees_table partial contains conditional logic so that it's only
displayed if there are services available in marriage_abroad_services.yml.
There are no services for ceremonies in the Czech Republic so the outer
conditional in this template is redundant.

Removing the conditional will make it easier to switch to rendering the
os_affirmation/fees partial in a later commit.