Remove link_to_consular_fees from no cni outcome
The link_to_consular_fees partial only renders content if there are services
available in the chosen ceremony country. None of the ceremony countries that
lead to this outcome have data in marriage_abroad_services.yml which means that
rendering link_to_consular_fees has no effect.

I used `script/marriage-abroad-countries-leading-to-outcome.rb` (with _all_
countries added to the marriage-abroad regression tests) to determine the
following countries that lead to this outcome:

* afghanistan
* american-samoa
* andorra
* aruba
* benin
* bhutan
* bonaire-st-eustatius-saba
* burkina-faso
* burundi
* cape-verde
* central-african-republic
* chad
* comoros
* congo
* curacao
* djibouti
* equatorial-guinea
* eritrea
* gabon
* guinea
* guinea-bissau
* haiti
* iraq
* liechtenstein
* mali
* marshall-islands
* mauritania
* micronesia
* niger
* palau
* paraguay
* saint-barthelemy
* san-marino
* sao-tome-and-principe
* south-sudan
* st-maarten
* st-martin
* suriname
* timor-leste
* togo
* western-sahara