Content for same sex marriage in Italy
Within contains, changes to the services and erb templates to support legalisation of same sex marriage in Italy.

Update to the fees definition for same sex marriage in Italy and other countries.

Also, reuse of the captured similarities (i.e added to content added to erb partials) from the previous commit.

I have run the regression tests and found changes to the existing regression test artefacts affecting the following.

/marriage-abroad/y/italy/ceremony_country/partner_british/same_sex
/marriage-abroad/y/italy/ceremony_country/partner_local/same_sex
/marriage-abroad/y/italy/ceremony_country/partner_other/same_sex
/marriage-abroad/y/italy/third_country/partner_british/same_sex
/marriage-abroad/y/italy/third_country/partner_local/same_sex
/marriage-abroad/y/italy/third_country/partner_other/same_sex
/marriage-abroad/y/italy/uk/partner_british/same_sex
/marriage-abroad/y/italy/uk/partner_local/same_sex
/marriage-abroad/y/italy/uk/partner_other/same_sex