Regenerate regression tests artefacts (Marriage Abroad)
This commit shows the affected/changes to regression test artefacts as a result of the Content changes for same sex
marriage legalisation in Italy.

I have reviewed the changes to the regression test artefacts and it is as expected.

The extraction of similar text from the opposite sex outcome to partials that
have been shared with the same sex outcome, did not in any way change the
Content within it.

As a result, of the addition of the consular fees table options for ceremony and
third country, on running the regression tests no changes were recorded.

This is becuase the changes will only affect same sex marriages and not opposite sex.

On applying the created Italian partials and text from the document.

I ran the regresstion tests and at this point, apparant changes merged and recorded.

Already existing unit, regression and integration tests passed, giving assurance that the aforementioned commit/changes
have had no far reaching effect.