Add a function to calculate reduced maintenance loan
The reduced maintenance loan only applies to dental/medical/healthcare
students who are in their 5th of 6th year of study. It is different to 
the general maintenance loan that already exists.

Should refactor the REDUCED_MAINTENTANCE_LOAN_AMOUNTS into rates YAML 
file so the values can be easily updated as an uprating.