Stop relying on slugs in areas returned by the imminence api
Imminence is removing the slugs because they are based on transforming the
names, which aren't static pieces of data and so shouldn't be relied on.
The only real change here is how the benefit-cap-calculator determines if
the claimant is based in London or not.  Instead of looking for an area
with the slug 'london' we look for an area with the type EUR (European
parliament constituency) and the name 'London'.  This is no more robust
than checking against the slug, as the names of the constituencies could
change, or the boundaries could change, or we could leave Europe and no
longer need these areas, but slug is going away so we have to do
something.

A better solution might be to check for known london councils, or use
mapit directly instead of imminience and check for different area types
(GLA for example) that are more permanent and idicative of a postcode
being in London.  We may also need to refer to the legislation to see
how it defines what "in London" means as it may not equate to "votes in
the London European Parliament Consitituency".