Remove July's AB testing code.
This commit removes the start button AB testing code for the month of July.

This will in effect, stop the creation of cookies and allocation of users to cohorts respectively.

Affecting only the following smart answers:

1. Overseas passports
2. Calculate your child maintenance
3. Marriage abroad

After the removal of the files and require entry in the javascript manifest file,

I ran all the regression, unit and flow tests and I did not record any errors.