Update transit (to Ireland ) for Electronic Visa Waiver Group
In this commit the focus is on Scenario H in the doc[1], which is for the transit to The Republic of Ireland.

This commit contains content and logic for rendering the changes made to the transit route (to Ireland ) for
Electronic Visa Waiver Group countries/passport holders with the intention of visiting the UK on transit to a
Common Travel Area (i.e CTA).

There's no need to update the unit or flow tests as the changes made here were mostly content changes
as recommended in the google doc [1].

[1] https://docs.google.com/document/d/12PnaBt8B797xibETodkUdNTaBTMqN5yfWbvt275XAUU/edit?ts=57bc67cd#