Move end of sentence from partial back into template
This moves a partial in calculate-statutory-sick-pay into the template
as it is only used that single time and therefore not necessary
and as it is inserted at the end of a sentence which is
not a good practice as it breaking up things which belong together.

This is to prepare for linking all partials to GitHub on Integration.
That would need to wrap each partial in a `div`
which would create invalid HTML in this case
and might lead to unexpected browser behaviour.

The substitute code for the partial is not inserted in a new line
because that would mean whitespace changes in 1260 artefacts.

The changes appear e.g. on this outcome page:
/calculate-statutory-sick-pay/y/none/yes/no/2016-08-07/2016-08-23/no/eight_weeks_more/monthly/2016-07-30/2016-05-30/1200.0/1,2,3,4,5