Move items of same list together
This moves a list item from a help-if-you-are-arrested-abroad outcome
from after a partial to the end of that partial
which includes the rest of the list.
This is to not break something apart which logically belongs together.

This is to prepare for linking all partials to GitHub on Integration.
That would need to wrap each partial in a `div`
which would create invalid HTML in this case
and might lead to unexpected browser behaviour.

An example outcome to test this on is:
/help-if-you-are-arrested-abroad/y/afghanistan