Move table head to rest of table
This moves the start of a table within a marriage-abroad template
into the following partial containing the rest of the table.

This is to prepare for linking all partials to GitHub on Integration.
That would need to wrap each partial in a `div`
which would create invalid HTML in this case
and might lead to unexpected browser behaviour.

The table appears on most outcomes, e.g. on
/marriage-abroad/y/dominican-republic/uk/partner_british/same_sex