Move bullets of single list items outside of partials
This moves list item bullets from partials
which only contain that single list item
in maternity-paternity-calculator and energy-grants-calculator
into the template calling the partial.
That way it doesn't matter what's in the partial,
it will always be valid HTML (unless it's empty).

This is to prepare for linking all partials to GitHub on Integration.
That would need to wrap each partial in a `div`
which would create invalid HTML in this case
and might lead to unexpected browser behaviour.

Example outcomes which show some of the changes but not all of them:
* /energy-grants-calculator/y/all_help/benefits,permission/1970-01-01/child_tax_credit,income_support/before-1940/house/electric_heating,loft_attic_conversion,mains_gas
* /maternity-paternity-calculator/y/maternity/2016-11-13/yes/2016-11-02/no