Update question, response and expected results
As a result of recently requested changes that will be affecting Jamaica, Jordan and Zimbabwe

I have updated the overseas-passports-questions-and-responses.yml and
overseas-passports-responses-and-expected-results.yml files with the aforementioned countries.

This will allow for the generation of regression artefacts before and after the proposed  changes, showing
the difference after the changes were made.

This commit also shows the affected/changes to regression test artefacts as a result of the addition of
Jamaica, Jordan and Zimbabwe to the the overseas-passports-questions-and-responses.yml and
overseas-passports-responses-and-expected-results.yml files.

I have reviewed the changes to the regression test artefacts and it is as expected showing the current state.

Already existing unit, regression and integration tests passed, giving
assurance that the aforementioned changes have had no (interfering) affect to
other smart answers.