Amendment to Durham Passport Office address
The address of the Durham Passport Office has changed, so I've amended the _sending_your_application.govspeak.erb for <% when 'durham' %>

Needs to go live on 06/09/2016.

This commit also shows the affected/changes to regression test artefacts as a result change to the application address.

This changes affects the following countries:

* Benin
* Jamaica
* Jordan
* Nigeria
* South Sudan
* The Occupied Palestinian Territories (i.e Jerusalem or Westbank)
* Zimbabwe

I have reviewed the changes to the regression test artefacts and it is as
expected showing the changes the address only.

Already existing unit, regression and integration tests passed, giving
assurance that the aforementioned changes have had no (interfering) affect to
other smart answers.