"Getting your passport" content changes for overseas passports
* Update collection office for Jamaica, Jordan and Zimbabwe

This commit also shows the affected/changes to regression test artefacts as a result change to the "Getting your passport" section.

This changes affects the following countries:

* Jamaica
* Jordan
* Zimbabwe

I have reviewed the changes to the regression test artefacts and it is as
expected showing the changes the delivery information only.

Already existing unit, regression and integration tests passed, giving
assurance that the aforementioned changes have had no (interfering) affect to
other smart answers.