Generate regression tests artefacts (SEC)
This commit shows the affected/changes to regression test artefacts as a result
of the changes (i.e addition of two new options for new and used cars) from
previous commits.

I have reviewed the changes to the regression test artefacts and it is as
expected.

This commit composes of two parts, the first half create new artefacts for the
new and used cars and the other half comprises of deleted artefacts from the
removed yes option.

At this point existing unit, regression and integration tests passed, giving
assurance that the aforementioned changes have had no (interfering) affect to
other smart answers.