Update questions, responses and expected results (LIC)
This commit contains an update to the questions, responses and expected result for
the Landlord Immigration Check smart answers.

This update accounts for the EU/EEA route of this smart answer only.

The following node have been removed and in effect no longer being exercised
- :has_certificate?
- :has_documents?
- :has_residence_card_or_eu_eea_swiss_family_member?
- :has_uk_passport?
- :outcome_can_rent_but_check_will_be_needed_again
- :outcome_can_rent_for_12_months
- :right_to_abode?
- :time_limited_to_remain?

At the moment it is not clear if the aforementioned will play a part in the future.