Hard code embassy addresses for same-sex Japan
Hard code the British Embassy and Consulate addresses in the same-sex
outcome for Japan.

Usually, embassy addresses are retrieved from Whitehall. A shared
template in overseas passports then loops through these addresses and
then renders them.

The problem we have is that the Consulate in Osaka no longer provides
marriage services. However, as it is still valid Consulate we can’t
remove it from Whitehall, but we need a way to no longer display it on 
the marriage-abroad same-sex outcome for Japan.

The simplest solution is to hard code the addresses for both the Embassy
and Consulate in Japan, prove that the test artefacts have not been 
affected, and then remove the Consulate address in the following
commit.