Regenerate regression test artefacts (MA)
This commit shows the affected/changes to regression test artefacts as a
result of the content changes made in the previous commit.

Based on the above, I have reviewed the changes to the regression test
artefacts and they appear as expected.

With the aforementioned only affecting opposite and same sex outcomes
except where partner is not a Vietnam national.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.