Rake task to add smart answers to the search index
Previously, panopticon would send artifacts to Rummager. That flow is now being
deprecated in favour of a simpler approach which is to send documents directly
to rummager.

This commit adds a class that allows us to add smart answers to the
search index directly. It also adds a rake task `rummager:index_all` whose job
is to index smart answers in Rummager.

This feature requires the following in place before it can be merged:
- [ ] disable https://github.com/alphagov/panopticon/blob/master/app/models/enhancements/artefact.rb#L17
- [ ] add this rake task to https://github.com/alphagov/govuk-app-deployment/blob/ccbebcbc02aacd19e94f82659735f88cd72cdf24/smartanswers/config/deploy.rb#L22

The payload now includes the following extra attributes:

- publishing_app
- rendering_app
- content_id