Use smart answer name to run regression tests
Given the lengthy period it takes to run the regression tests within this repo.
It appears that the jenkins.sh script has omitted running the regression tests.

Hence this commit seeks to include the regression test as part of the jenkins script.

This is reliant on the git branch being setup with the name of the smart-answer first followed by a
forward slash and the the actually name of the branch (i.e some git work flow conventions have referred to this as a feature branching pattern).

This convention in effect provides the basis for sensible assumption to be made, that the part
before the forward slash is indeed the name of the smart answer flow (i.e ideal scenario).

This is then used to provide RUN_REGRESSION_TESTS environment variable with the required.

if the branch is named without the forward slash, the branch name is returned preceded by
asterisk (*) from the parse_git_branch function.

if the value return from the parse_git_branch function does not match a smart answer flow

The aforementioned (i.e edge cases) will not invoke or run any regression tests (i.e gracefully).