Change logic to render Green Deal callout
Affected outcome:
outcome_bills_and_measures_on_benefits_eco_eligible.govspeak.erb

If the user is not eligible for a winter fuel or cold weather payment,
but is eligible for the deal, the green deal title will be displayed.

In this circumstance, it is ok to display the green deal callout
directly below the title.

If the user is eligible for either a winter fuel payment or a cold
weather payment (i.e. help from the government), the callout text should
be displayed after the section on the help from the government.

There will be a future commit that displays a Green deal sub heading
above the green deal callout if users are eligible for a winter fuel or
cold weather payment.

Regenerate test artefacts:

The green callout text moves as expected below the winter fuel payment or
cold weather payment bullet if the user is eligible for either.